New structure and refactor
