+++ frontend/locales/en/fishbowl.json
@@ -114,6 +114,7 @@
         "insightful": "Insightful"
     },
     "prefishbowl": {
-        "onBoardingHelp": "Don't know how a fishbowl works? Click here"
+        "onBoardingHelp": "Don't know how a fishbowl works? Click here",
+        "connectWithUsers": "To connect with the people in the room through your social networks, sign up on Stooa."
     }
 }
