+++ frontend/locales/fr/fishbowl.json
@@ -106,6 +106,7 @@
         "insightful": "Intéressant"
     },
     "prefishbowl": {
-        "onBoardingHelp": "Don't know how a fishbowl works? Click here"
+        "onBoardingHelp": "Don't know how a fishbowl works? Click here",
+        "connectWithUsers": "To connect with the people in the room through your social networks, sign up on Stooa."
     }
 }
