+++ frontend/src/components/App/Onboarding/index.tsx
@@ -7,234 +7,43 @@
  * file that was distributed with this source code.
  */
 
-import React, { useState, useEffect, useRef } from 'react';
+import React from 'react';
 import useTranslation from 'next-translate/useTranslation';
-import Trans from 'next-translate/Trans';
-import SlickSlider from 'react-slick';
-import { gsap } from 'gsap';
 
-import ModalOnBoarding from '@/components/App/ModalOnBoarding';
-import { getOnBoardingCookie, setOnBoardingCookie } from '@/lib/auth';
-import { pushEventDataLayer } from '@/lib/analytics';
-
-import ArrowPrev from '@/ui/svg/arrow-prev.svg';
-import ArrowNext from '@/ui/svg/arrow-next.svg';
 import Cross from '@/ui/svg/cross.svg';
 import QuestionMark from '@/ui/svg/questionmark.svg';
-import OnboardingWrapper, {
-  Icon,
-  Slider,
-  Slide,
-  Tooltip
-} from '@/components/App/Onboarding/styles';
-import onBoardingDataWithIntroduction from '@/components/App/Onboarding/dataWithIntroduction.json';
-import onBoardingDataWithoutIntroduction from '@/components/App/Onboarding/dataWithoutIntroduction.json';
+import OnboardingWrapper, { Icon, Tooltip } from '@/components/App/OnBoarding/styles';
 import { useStooa } from '@/contexts/StooaManager';
-import Button from '@/components/Common/Button';
 
-interface Props {
-  initialized: boolean;
-  isModerator: boolean;
-}
-
-interface Arrow {
-  currentSlide?: number;
-  slideCount?: number;
-  children?: React.ReactElement;
-}
-
-const SlickButtonFix = ({ children, ...props }) => (
-  <button className="arrow prev" {...props}>
-    {children}
-  </button>
-);
-
-// This props are required to fix the issue with the SlickSlider component
-// ref: https://github.com/akiran/react-slick/issues/1195
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-const PrevArrow = ({ currentSlide, slideCount, children, ...props }: Arrow) => (
-  <SlickButtonFix {...props}>
-    <ArrowPrev />
-  </SlickButtonFix>
-);
-
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-const NextArrow = ({ currentSlide, slideCount, children, ...props }: Arrow) => (
-  <SlickButtonFix {...props}>
-    <ArrowNext />
-  </SlickButtonFix>
-);
-
-const Onboarding: React.FC<Props> = ({ isModerator }) => {
-  const sliderRef = useRef(null);
-  const [active, setActive] = useState(false);
-  const [alreadySeen, setAlreadySeen] = useState(false);
-  const [activeTooltip, setActiveTooltip] = useState(false);
-  const { data: fishbowlData } = useStooa();
-  const [data, setData] = useState([]);
+const OnBoardingButton = () => {
+  const {
+    activeOnBoardingTooltip,
+    setActiveOnBoardingTooltip,
+    toggleOnBoarding,
+    setOnBoardingTooltipSeen
+  } = useStooa();
   const { t } = useTranslation('on-boarding');
-
-  const SLIDE_SETTINGS = {
-    dots: true,
-    speed: 500,
-    fade: true,
-    infinite: false,
-    lazyload: true,
-    slidesToShow: 1,
-    slidesToScroll: 1,
-    focusOnSelect: false,
-    prevArrow: <PrevArrow />,
-    nextArrow: <NextArrow />,
-    beforeChange: (current: number) => {
-      clearAnimation(current);
-      hideSlide(current, 0.5);
-    },
-    afterChange: (current: number) => {
-      showSlide(current);
-    }
-  };
-
-  const toggleOnBoarding = () => {
-    pushEventDataLayer({
-      action: active ? 'OnBoarding close' : 'OnBoarding open',
-      category: 'Header',
-      label: window.location.href
-    });
-
-    setActive(!active);
-  };
-
-  const skipOnBoarding = () => {
-    toggleOnBoarding();
-    setOnBoardingCookie(isModerator);
-    if (!alreadySeen) setActiveTooltip(true);
-  };
-
-  const shouldShow = () => {
-    const cookie = getOnBoardingCookie(isModerator);
-
-    if (!cookie) {
-      setActive(true);
-      setAlreadySeen(false);
-    }
-  };
-
-  const clearAnimation = (slide: number) => {
-    const items = document.querySelectorAll(`.slide[data-slide="${slide}"] .animate`);
-
-    items.forEach(item => {
-      item.removeAttribute('style');
-    });
-  };
-
-  const hideSlide = (slide: number, delay = 0) => {
-    const items = document.querySelectorAll(`.slide[data-slide="${slide}"] .animate`);
-
-    items.forEach(item => {
-      gsap.to(item, 0, {
-        opacity: 0,
-        y: 30,
-        ease: 'none',
-        delay
-      });
-    });
-  };
-
-  const showSlide = (slide: number, delay = 0.3) => {
-    const items = document.querySelectorAll(`.slide[data-slide="${slide}"] .animate`);
-
-    items.forEach((item, i) => {
-      gsap.to(item, 1, {
-        opacity: 1,
-        y: 0,
-        ease: 'Power3.easeOut',
-        delay: delay + 0.1 * i
-      });
-    });
-  };
-
-  useEffect(() => {
-    if (fishbowlData.hasIntroduction) {
-      setData(onBoardingDataWithIntroduction[isModerator ? 'moderator' : 'participant']);
-    } else {
-      setData(onBoardingDataWithoutIntroduction[isModerator ? 'moderator' : 'participant']);
-    }
-
-    shouldShow();
-  }, []); // eslint-disable-line react-hooks/exhaustive-deps
-
-  useEffect(() => {
-    if (!active) sliderRef.current.slickGoTo(0);
-    else showSlide(0);
-  }, [active]);
 
   return (
     <OnboardingWrapper>
-      <Icon onClick={toggleOnBoarding}>
+      <Icon onClick={() => toggleOnBoarding('header')}>
         <QuestionMark />
       </Icon>
-      {activeTooltip && (
+      {activeOnBoardingTooltip && (
         <Tooltip className="info">
           <p>{t('tooltip')}</p>
           <button
             onClick={() => {
-              setActiveTooltip(false);
-              setAlreadySeen(true);
+              setActiveOnBoardingTooltip(false);
+              setOnBoardingTooltipSeen(true);
             }}
           >
             <Cross />
           </button>
         </Tooltip>
       )}
-      <ModalOnBoarding closeModal={skipOnBoarding} show={active}>
-        <Slider>
-          <SlickSlider {...SLIDE_SETTINGS} ref={sliderRef}>
-            {data.map((item, i) => (
-              <Slide
-                key={`slide-${i}`}
-                className={`${isModerator ? 'moderator' : 'participant'} slide slide-${i}`}
-                data-slide={i}
-              >
-                <div className="left">
-                  <div className="animate">
-                    {item.pretitle && <p>{t(item.preTitle)}</p>}
-                    {item.title && <h2 className="title-md">{t(item.title)}</h2>}
-                    {item.text && (
-                      <Trans
-                        i18nKey={item.text}
-                        components={{
-                          div: <div />,
-                          p: <p />,
-                          ul: <ul />,
-                          li: <li />,
-                          strong: <strong />
-                        }}
-                      />
-                    )}
-                    {i === data.length - 1 ? (
-                      <Button size="medium" onClick={skipOnBoarding}>
-                        {t('done')}
-                      </Button>
-                    ) : (
-                      <Button variant="subtleLink" onClick={skipOnBoarding}>
-                        {t('skip')}
-                      </Button>
-                    )}
-                  </div>
-                </div>
-                <div className="right">
-                  {item.img1 && <img className="animate img-1" src={item.img1} alt="" />}{' '}
-                  {/* eslint-disable-line @next/next/no-img-element */}
-                  {item.img2 && <img className="animate img-2" src={item.img2} alt="" />}{' '}
-                  {/* eslint-disable-line @next/next/no-img-element */}
-                </div>
-              </Slide>
-            ))}
-          </SlickSlider>
-        </Slider>
-      </ModalOnBoarding>
     </OnboardingWrapper>
   );
 };
 
-export default Onboarding;
+export default OnBoardingButton;
